prevent signed overflow

+++ MagickCore/property.c
@@ -1450,7 +1450,7 @@ static void GetEXIFProperty(const Image *image,const char *property,
           dir_offset=(ssize_t) ReadPropertySignedLong(endian,q+8);
           if ((dir_offset < 0) || (size_t) dir_offset >= length)
             continue;
-          if ((dir_offset+(ssize_t) number_bytes) < dir_offset)
+          if (((size_t) dir_offset+number_bytes) < (size_t) dir_offset)
             continue;  /* prevent overflow */
           if ((size_t) (dir_offset+(ssize_t) number_bytes) > length)
             continue;
